# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get Me",
        "operationId": "get_me_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithProfileDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/{user_id}": {
      "get": {
        "summary": "Get By Id",
        "operationId": "get_by_id_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithProfileDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/travel_posts/": {
      "get": {
        "summary": "Get All",
        "operationId": "get_all_travel_posts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TravelPostDTO"
                  },
                  "type": "array",
                  "title": "Response Get All Travel Posts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create",
        "operationId": "create_travel_posts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTravelPostDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPostDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/travel_posts/{post_id}": {
      "get": {
        "summary": "Get By Id",
        "operationId": "get_by_id_travel_posts__post_id__get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPostDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update",
        "operationId": "update_travel_posts__post_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTravelPostDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPostDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccessTokenDTO": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "AccessTokenDTO"
      },
      "CreateTravelPostDTO": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cost": {
            "type": "integer",
            "title": "Cost"
          },
          "transportation_rating": {
            "type": "integer",
            "title": "Transportation Rating"
          },
          "safety_rating": {
            "type": "integer",
            "title": "Safety Rating"
          },
          "overcrowding_rating": {
            "type": "integer",
            "title": "Overcrowding Rating"
          },
          "nature_rating": {
            "type": "integer",
            "title": "Nature Rating"
          },
          "cultural_heritage_sites": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cultural Heritage Sites"
          }
        },
        "type": "object",
        "required": [
          "title",
          "cost",
          "transportation_rating",
          "safety_rating",
          "overcrowding_rating",
          "nature_rating",
          "cultural_heritage_sites"
        ],
        "title": "CreateTravelPostDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginUserDTO": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginUserDTO"
      },
      "TravelPostDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cost": {
            "type": "integer",
            "title": "Cost"
          },
          "transportation_rating": {
            "type": "integer",
            "title": "Transportation Rating"
          },
          "safety_rating": {
            "type": "integer",
            "title": "Safety Rating"
          },
          "overcrowding_rating": {
            "type": "integer",
            "title": "Overcrowding Rating"
          },
          "nature_rating": {
            "type": "integer",
            "title": "Nature Rating"
          },
          "cultural_heritage_sites": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cultural Heritage Sites"
          },
          "user": {
            "$ref": "#/components/schemas/UserDTO"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "cost",
          "transportation_rating",
          "safety_rating",
          "overcrowding_rating",
          "nature_rating",
          "cultural_heritage_sites",
          "user"
        ],
        "title": "TravelPostDTO"
      },
      "UserDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "UserDTO"
      },
      "UserWithProfileDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "travel_posts": {
            "items": {
              "$ref": "#/components/schemas/TravelPostDTO"
            },
            "type": "array",
            "title": "Travel Posts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "travel_posts"
        ],
        "title": "UserWithProfileDTO"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}